requirements.txt
Django==1.4
South==0.7.5
dj-database-url==0.2.0
django-piston==0.2.3
psycopg2==2.4.5
wsgiref==0.1.2
